+++ stuff/runner.h
@@ -2,7 +2,7 @@ struct runner;
 
 struct runner_tick {
 	struct fcsim_block_stat *stats;
-	int index;
+	uint64_t index;
 };
 
 struct runner *runner_create(void);
@@ -15,6 +15,6 @@ void runner_stop(struct runner *runner);
 
 void runner_get_tick(struct runner *runner, struct runner_tick *tick);
 
-int runner_get_won_tick(struct runner *runner);
+uint64_t runner_get_won_tick(struct runner *runner);
 
 void runner_set_frame_limit(struct runner *runner, uint64_t frame_limit_us);
